+++ src/useDocumentVisibility.ts
@@ -1,12 +1,10 @@
 import { useEffect, useState } from 'react';
 import { canUseDOM, managedEventListener } from './utils';
 
-type StandardVisibilityState = Exclude<VisibilityState, 'prerender'>;
-
 /**
  * Tracks visibility of the page.
  *
- * @returns {VisibilityState} Visibility state of the [`document`](https://developer.mozilla.org/docs/Web/API/Document), which is `'visible'` by default.
+ * @returns Visibility state of the [`document`](https://developer.mozilla.org/docs/Web/API/Document), which is `'visible'` by default.
  *
  * @example
  * function Example() {
@@ -17,17 +15,15 @@ type StandardVisibilityState = Exclude<VisibilityState, 'prerender'>;
  *   // ...
  * }
  */
-export default function useDocumentVisibility() {
+export default function useDocumentVisibility(): VisibilityState {
   const [visibility, setVisibility] = useState(
-    canUseDOM
-      ? (document.visibilityState as StandardVisibilityState)
-      : 'visible', // TODO: Consider using 'prerender'
+    canUseDOM ? document.visibilityState : 'visible', // TODO: Consider using 'prerender'
   );
 
   useEffect(
     () =>
       managedEventListener(document, 'visibilitychange', () => {
-        setVisibility(document.visibilityState as StandardVisibilityState);
+        setVisibility(document.visibilityState);
       }),
     [],
   );